Moved Lti1 to Lti\v1
