+++ LtiLibrary.NetCore.Tests/ContentItemTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using LtiLibrary.NetCore.ContentItems;
-using LtiLibrary.NetCore.Lti1;
+using LtiLibrary.NetCore.Lti.v1;
 using LtiLibrary.NetCore.Tests.SimpleHelpers;
 using Xunit;
