+++ LtiLibrary.NetCore/Lti/v1/IBasicLaunchRequest.cs
@@ -1,4 +1,4 @@
-namespace LtiLibrary.NetCore.Lti1
+namespace LtiLibrary.NetCore.Lti.v1
 {
     internal interface IBasicLaunchRequest : ILtiRequest
     {
