+++ LtiLibrary.NetCore/Outcomes/v2/OutcomesClient.cs
@@ -8,7 +8,7 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Security.Cryptography;
 using LtiLibrary.NetCore.Lis.v2;
-using LtiLibrary.NetCore.Lti1;
+using LtiLibrary.NetCore.Lti.v1;
 
 namespace LtiLibrary.NetCore.Outcomes.v2
 {
